requirements.txt
streamlit
openai==0.28
langchain
unstructured
pdf2image
pdfminer
streamlit-chat
pypdfium2
chromadb
tiktoken
python-dotenv
requests
pydub
gtts
certifi
git+https://github.com/openai/whisper.git
streamlit-audiorecorder
whisper